added loader and conditions to update todos

+++ app/ui/Todo.tsx
@@ -4,6 +4,8 @@ import axios from "axios";
 import { API } from "../lib/constants";
 import { useDispatch, useSelector } from "react-redux";
 import { removeTodo, TodoStateInterface, updateTodo } from "../GlobalRedux/Features/todos/todoSlice";
+import Image from "next/image";
+import Loader from "./Loader";
 
 interface Props {
   todo: TodoType;
@@ -12,58 +14,82 @@ interface Props {
 export default function Todo({ todo }: Props) {
   const dispatch = useDispatch();
   const [title, setTitle] = useState(todo.title);
+  const [prevTitle, setPrevTitle] = useState(todo.title);
   const [isImportant, setIsImportant] = useState(false);
-  const todos = useSelector((state: { todoReducer: TodoStateInterface }) => state.todoReducer.todos);
-
+  const [isRequest, setIsRequest] = useState(false);
 
   function updatingTodo(completedValue?: boolean) {
+    if (isRequest) return;
+    
     const updatedTodo = {
       ...todo,
       title,
       completed: completedValue !== undefined ? completedValue : todo.completed,
     };
-
+    
+    setIsRequest(true);
     axios.patch(`${API}/todos/${todo.id}`, {
       title, completed: updatedTodo.completed,
     })
       .then(res => {
         dispatch(updateTodo(updatedTodo));
+        setIsRequest(false);
       })
       .catch(err => {
+        setIsRequest(false);
         throw new Error(err);
       });
   }
 
   function deleteTodo() {
+    if (isRequest) return;
+
+    setIsRequest(true);
     axios.delete(`${API}/todos/${todo.id}`)
       .then(res => {
         dispatch(removeTodo(todo.id));
+        setIsRequest(false);
       })
       .catch(err => {
+        setIsRequest(false);
         throw new Error(err);
       });
   }
 
   return (
-    <div className={`mb-0.5 p-3 flex bg-neutral-800 ${todo.completed ? 'text-slate-500' : 'text-slate-100'}  rounded justify-between`}>
+    <div className={`mb-0.5 p-3 flex bg-neutral-800 ${todo.completed ? 'text-slate-500' : 'text-slate-100'} rounded justify-between relative`}>
       <div className={`flex grow ${todo.completed ? 'line-through' : ''}`}>
+        {isRequest && <Loader className="absolute left-1.5" />}
         <input
           type="checkbox"
           className="mr-2 self-start mt-1.5"
           checked={todo.completed}
-          onChange={() => updatingTodo(!todo.completed)}
+          onChange={() => {
+            if (!isRequest) updatingTodo(!todo.completed);
+          }}
         />
 
         <input
           type="text"
           value={title}
-          onChange={(e) => setTitle(e.target.value)}
+          onFocus={() => {
+            setPrevTitle(title);
+          }}
+          onChange={(e) => {
+            if (!isRequest) {
+              setTitle(e.target.value);
+            }
+          }}
           className="w-full mx-3 bg-transparent outline-none resize-none"
           disabled={todo.completed}
           onKeyUp={(e) => {
-            if (e.key === 'Enter') updatingTodo();
+            if (e.key === 'Enter' && prevTitle !== title) {
+              updatingTodo();
+            }
           }}
-          onBlur={() => updatingTodo()}
+          onBlur={() => {
+            if (prevTitle !== title) updatingTodo();
+          }}
         />
       </div>
 
@@ -72,7 +98,9 @@ export default function Todo({ todo }: Props) {
           type="checkbox"
           className=""
           checked={isImportant}
-          onChange={() => setIsImportant(!isImportant)}
+          onChange={() => {
+            if (!isRequest) setIsImportant(!isImportant);
+          }}
         />
 
         {todo.completed &&
